action checkTerms, initTerms on app.js
