+++ routers/api/terms.js
@@ -1,30 +1,9 @@
 const router = require('express').Router();
-const { initTerms } = require('../../services/terms')
-const { insertTerms } = require('../../database/actions/terms')
+// const { initTerms } = require('../../services/terms')
+// const { insertTerms } = require('../../database/actions/terms')
 
 
-router.get('/getTerms', (req,res) => {
-    const { page, size } = req.query
-     
-
-    initTerms({ page, size })
-        .then(({data}) => {
-            const terms = data._embedded.terms
-
-            const termsData = terms.map(term => ({
-                key: term.obo_id,
-                label: term.label,
-                synonyms: term.synonyms ? term.synonyms.join(', ') : '-',
-                obo_id: term.obo_id,
-                term_editor: term.annotation['term editor']
-                 ? term.annotation['term editor'].join(', ')
-                 : '-',
-                has_children: term.has_children,
-            }))
-            insertTerms(termsData)
-        })
-        .catch(err => res.status(400).send(err))
-})
+// router.get('/getterms', (req,res) => { })
 
 /* router.put('/updateTerm', (req, res) => {
